requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
docopt==0.6.2
google-api-core==1.25.1
google-auth==1.24.0
google-cloud-secret-manager==2.2.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.35.0
idna==2.10
importanize==0.7.0
Jinja2==2.11.2
jsonpath-ng==1.5.2
kubernetes==12.0.1
libcst==0.3.16
MarkupSafe==1.1.1
mergedeep==1.3.1
more-itertools==8.6.0
mypy-extensions==0.4.3
nested-lookup==0.2.21
oauthlib==3.1.0
pathlib2==2.3.5
ply==3.11
progress==1.5
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pykwalify==1.8.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
ruamel.yaml==0.16.12
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
unflatten==0.1
urllib3==1.26.3
websocket-client==0.57.0
